Final Update
This is my final Update. Some of my past commits did not have explanations because I was trouble shooting issues. Hopefully this will solve them.

Changes: The path to the data set should be different because I deployed the app via github and the pathway notation is different there. I.E the streamlit app pulled from github and not my computer

Note to Professor: I don't know how to go back and change past commits, I hope this is detailed enough. Let me know if there is an issues